+++ application_list.py
@@ -1,15 +1,15 @@
 application_list = [
     {
-    'route': 'investor',
-    'image_path': 'pictures/spam.jpg',
-    'title': 'Spam Detection',
-    'project_name': ' Automatic Spam Filtering',
-    'project_description': 'Automatic spam filtering is a technology that employs algorithms to analyze incoming emails, messages, or content and identify and categorize unwanted or malicious spam items. By examining various attributes like sender, content, and patterns, it helps users manage their communications more effectively and protect against unsolicited or harmful messages.',
-    'project_demo_link': 'http://aipocdemo.creatingwow.in/',
-    'project_demo_video': 'https://drive.google.com/file/d/1aGaoqv786_944dC0h6oZDDfTaTojbiMI/view?t=10s',
-    'project_demo_ppt': 'https://docs.google.com/presentation/d/15l41H5aua3hu08ZuFjAg6jgLW0iNEr7Z/edit#slide=id.g25e105483f5_0_832',
-    'project_model_code_link': 'https://github.com/rahuldhamecha-infobeans/spam-ham-detection/blob/main/jupyter-notbook-code/Pragati-spam-detection.ipynb',
-    'created_by': 'Prgati Nande'
+        'route': 'spamham.investor',
+        'image_path': 'pictures/spam.jpg',
+        'title': 'Spam Detection',
+        'project_name': ' Automatic Spam Filtering',
+        'project_description': 'Automatic spam filtering is a technology that employs algorithms to analyze incoming emails, messages, or content and identify and categorize unwanted or malicious spam items. By examining various attributes like sender, content, and patterns, it helps users manage their communications more effectively and protect against unsolicited or harmful messages.',
+        'project_demo_link': 'http://aipocdemo.creatingwow.in/',
+        'project_demo_video': 'https://drive.google.com/file/d/1aGaoqv786_944dC0h6oZDDfTaTojbiMI/view?t=10s',
+        'project_demo_ppt': 'https://docs.google.com/presentation/d/15l41H5aua3hu08ZuFjAg6jgLW0iNEr7Z/edit#slide=id.g25e105483f5_0_832',
+        'project_model_code_link': 'https://github.com/rahuldhamecha-infobeans/spam-ham-detection/blob/main/jupyter-notbook-code/Pragati-spam-detection.ipynb',
+        'created_by': 'Prgati Nande'
     },
     # {
     #         'route': 'spam_email',
@@ -37,30 +37,30 @@ application_list = [
     #     'created_by': 'Prgati Nande'
     # },
     {
-    'route': 'object_detection',
-    'image_path': 'pictures/object_detection.png',
-    'title': 'Live Object <span>Detection</span>',
-    # New Data
-    'project_name': 'Live Object Detection',
-    'project_description': 'Object detection is an important computer vision task used to detect instances of visual objects of certain classes (for example, humans, animals, cars, or buildings) in digital images such as photos or video frames. The goal of object detection is to develop computational models that provide the most fundamental information needed.',
-    'project_demo_link': 'http://aipocdemo.creatingwow.in/',
-    'project_demo_video': 'https://www.awesomescreenshot.com/video/20199304?key=ce8805adc24d99d9ce1056beb767b899',
-    'project_demo_ppt': 'https://docs.google.com/presentation/d/1Er7I-SuYBdED1ZGxow_gBhKmeqCzG4cs/edit#slide=id.g275375f36b8_6_0',
-    'project_model_code_link': 'https://github.com/rahuldhamecha-infobeans/spam-ham-detection/blob/main/jupyter-notbook-code/object_detection.py',
-    'created_by': 'Priyanka Bhandare'
+        'route': 'objectdetection.object_detection',
+        'image_path': 'pictures/object_detection.png',
+        'title': 'Live Object <span>Detection</span>',
+        # New Data
+        'project_name': 'Live Object Detection',
+        'project_description': 'Object detection is an important computer vision task used to detect instances of visual objects of certain classes (for example, humans, animals, cars, or buildings) in digital images such as photos or video frames. The goal of object detection is to develop computational models that provide the most fundamental information needed.',
+        'project_demo_link': 'http://aipocdemo.creatingwow.in/',
+        'project_demo_video': 'https://www.awesomescreenshot.com/video/20199304?key=ce8805adc24d99d9ce1056beb767b899',
+        'project_demo_ppt': 'https://docs.google.com/presentation/d/1Er7I-SuYBdED1ZGxow_gBhKmeqCzG4cs/edit#slide=id.g275375f36b8_6_0',
+        'project_model_code_link': 'https://github.com/rahuldhamecha-infobeans/spam-ham-detection/blob/main/jupyter-notbook-code/object_detection.py',
+        'created_by': 'Priyanka Bhandare'
     },
     {
-    'route': 'image_classification',
-    'image_path': 'pictures/spam.jpg',
-    'title': 'Image <span>Classifiaction</span>',
-    # New Data
-    'project_name': 'Image Classification using CNN',
-    'project_description': 'An image classification problem involves the task of categorizing images into predefined classes or categories. The goal is to develop a machine learning model that can accurately predict the class label of an image based on its visual content. In the field of biology and ecology, researchers often need to identify and classify various species of plants and animals for conservation efforts, ecosystem monitoring, and biodiversity studies.',
-    'project_demo_link': 'http://aipocdemo.creatingwow.in/',
-    'project_demo_video': 'https://www.loom.com/share/3150261db541434e896fc5101f6a5d6a',
-    'project_demo_ppt': 'https://docs.google.com/presentation/d/1SsfHdfOKd6KT1eWrI0mmhviL55OQe0hJ/edit?usp=drive_link&ouid=111877127295039292444&rtpof=true&sd=true',
-    'project_model_code_link': 'https://github.com/rahuldhamecha-infobeans/spam-ham-detection/blob/main/jupyter-notbook-code/image_classification.ipynb',
-    'created_by': 'Sharad Kumar Gupta & Aman Goel'
+        'route': 'imageclassification.image_classification',
+        'image_path': 'pictures/spam.jpg',
+        'title': 'Image <span>Classifiaction</span>',
+        # New Data
+        'project_name': 'Image Classification using CNN',
+        'project_description': 'An image classification problem involves the task of categorizing images into predefined classes or categories. The goal is to develop a machine learning model that can accurately predict the class label of an image based on its visual content. In the field of biology and ecology, researchers often need to identify and classify various species of plants and animals for conservation efforts, ecosystem monitoring, and biodiversity studies.',
+        'project_demo_link': 'http://aipocdemo.creatingwow.in/',
+        'project_demo_video': 'https://www.loom.com/share/3150261db541434e896fc5101f6a5d6a',
+        'project_demo_ppt': 'https://docs.google.com/presentation/d/1SsfHdfOKd6KT1eWrI0mmhviL55OQe0hJ/edit?usp=drive_link&ouid=111877127295039292444&rtpof=true&sd=true',
+        'project_model_code_link': 'https://github.com/rahuldhamecha-infobeans/spam-ham-detection/blob/main/jupyter-notbook-code/image_classification.ipynb',
+        'created_by': 'Sharad Kumar Gupta & Aman Goel'
     },
 
     # {
@@ -77,45 +77,45 @@ application_list = [
     #     'created_by': 'Prgati Nande'
     # },
     {
-    'route': 'facial_emotion_based_recommendation',
-    'image_path': 'pictures/songrec.jpg',
-    'title': 'Song Recommendations <br><span>Facial Emotion Based</span>',
-    # New Data
-    'project_name': 'Song Recommendations Facial Based Emotion',
-    'project_description': "The Facial Emotions-based Song Recommendation project aims to enhance user experience by leveraging facial emotion recognition to curate and recommend songs that align with the user's emotional state. Recognizing human emotions through facial expressions has gained significant traction, presenting an exciting opportunity for innovative applications. This project addresses the need for a novel approach to music recommendation by incorporating real-time facial emotion analysis. The core problem is to develop an intelligent system that can accurately detect the user's emotions from their facial expressions in real time.",
-    'project_demo_link': 'http://aipocdemo.creatingwow.in/',
-    'project_demo_video': 'https://drive.google.com/file/d/1lAdapKPxDOzdR-qa8PvNclD1VoOrDRla/view?usp=drive_link',
-    'project_demo_ppt': 'https://docs.google.com/presentation/d/1p8IbYQX3Wt39JyuJ0vYyU_McYs88mCGK/edit#slide=id.g2746efa11bc_5_8',
-    'project_model_code_link': 'https://github.com/rahuldhamecha-infobeans/spam-ham-detection/blob/main/jupyter-notbook-code/train.py',
-    'created_by': 'Sharad Kumar Gupta & Aman Goel'
+        'route': 'songsdetection.facial_emotion_based_recommendation',
+        'image_path': 'pictures/songrec.jpg',
+        'title': 'Song Recommendations <br><span>Facial Emotion Based</span>',
+        # New Data
+        'project_name': 'Song Recommendations Facial Based Emotion',
+        'project_description': "The Facial Emotions-based Song Recommendation project aims to enhance user experience by leveraging facial emotion recognition to curate and recommend songs that align with the user's emotional state. Recognizing human emotions through facial expressions has gained significant traction, presenting an exciting opportunity for innovative applications. This project addresses the need for a novel approach to music recommendation by incorporating real-time facial emotion analysis. The core problem is to develop an intelligent system that can accurately detect the user's emotions from their facial expressions in real time.",
+        'project_demo_link': 'http://aipocdemo.creatingwow.in/',
+        'project_demo_video': 'https://drive.google.com/file/d/1lAdapKPxDOzdR-qa8PvNclD1VoOrDRla/view?usp=drive_link',
+        'project_demo_ppt': 'https://docs.google.com/presentation/d/1p8IbYQX3Wt39JyuJ0vYyU_McYs88mCGK/edit#slide=id.g2746efa11bc_5_8',
+        'project_model_code_link': 'https://github.com/rahuldhamecha-infobeans/spam-ham-detection/blob/main/jupyter-notbook-code/train.py',
+        'created_by': 'Sharad Kumar Gupta & Aman Goel'
     },
     {
-    'route': 'mood_based_recommendation',
-    'image_path': 'pictures/songrec.jpg',
-    'title': 'Song Recommendations<br><span>Mood Based</span>',
-    # New Data
-    'project_name': 'Mood based song recommendation',
-    'project_description': 'Recommend the songs depends on entered song mood.',
-    'project_demo_link': 'http://aipocdemo.creatingwow.in/',
-    'project_demo_video': '#',
-    'project_demo_ppt': 'https://docs.google.com/presentation/d/1p8IbYQX3Wt39JyuJ0vYyU_McYs88mCGK/edit#slide=id.g2746efa11bc_5_8',
-    'project_model_code_link': 'https://github.com/rahuldhamecha-infobeans/spam-ham-detection/blob/main/jupyter-notbook-code/music-mood-classification.ipynb',
-    'created_by': 'Sharad Kumar Gupta & Aman Goel'
+        'route': 'songsdetection.mood_based_recommendation',
+        'image_path': 'pictures/songrec.jpg',
+        'title': 'Song Recommendations<br><span>Mood Based</span>',
+        # New Data
+        'project_name': 'Mood based song recommendation',
+        'project_description': 'Recommend the songs depends on entered song mood.',
+        'project_demo_link': 'http://aipocdemo.creatingwow.in/',
+        'project_demo_video': '#',
+        'project_demo_ppt': 'https://docs.google.com/presentation/d/1p8IbYQX3Wt39JyuJ0vYyU_McYs88mCGK/edit#slide=id.g2746efa11bc_5_8',
+        'project_model_code_link': 'https://github.com/rahuldhamecha-infobeans/spam-ham-detection/blob/main/jupyter-notbook-code/music-mood-classification.ipynb',
+        'created_by': 'Sharad Kumar Gupta & Aman Goel'
     },
 
-    {
-    'route': 'email_validator',
-    'image_path': 'pictures/spam.jpg',
-    'title': 'Email<br><span>Validator</span>',
-    # New Data
-    'project_name': 'Email Validator',
-    'project_description': 'The "Email Validator" typically involves a user or recipient failing to review or acknowledge their emails in a timely manner. This can result in missed communication, important messages going unnoticed, and potential disruptions in professional or personal interactions.',
-    'project_demo_link': 'http://aipocdemo.creatingwow.in/',
-    'project_demo_video': '#',
-    'project_demo_ppt': 'https://docs.google.com/presentation/d/166y33tUod1R_omgqPIeIwHTI-s-P8TFE/edit',
-    'project_model_code_link': 'https://github.com/rahuldhamecha-infobeans/spam-ham-detection',
-    'created_by': 'Prgati Nande & Sharad Kumar Gupta'
-    },
+    # {
+    #     'route': 'email_validator',
+    #     'image_path': 'pictures/spam.jpg',
+    #     'title': 'Email<br><span>Validator</span>',
+    #     # New Data
+    #     'project_name': 'Email Validator',
+    #     'project_description': 'The "Email Validator" typically involves a user or recipient failing to review or acknowledge their emails in a timely manner. This can result in missed communication, important messages going unnoticed, and potential disruptions in professional or personal interactions.',
+    #     'project_demo_link': 'http://aipocdemo.creatingwow.in/',
+    #     'project_demo_video': '#',
+    #     'project_demo_ppt': 'https://docs.google.com/presentation/d/166y33tUod1R_omgqPIeIwHTI-s-P8TFE/edit',
+    #     'project_model_code_link': 'https://github.com/rahuldhamecha-infobeans/spam-ham-detection',
+    #     'created_by': 'Prgati Nande & Sharad Kumar Gupta'
+    # },
 
     # {
     #     'route': 'multi_image_classification',
