+++ poc/imageclassification/templates/multi-image-classification.html
