+++ poc/songsdetection/templates/song-recommendation.html
@@ -1,10 +1,10 @@
-{% extends 'layouts/base.html' %} {% block page_title %} Mood Based Songs
+{% extends 'layouts/base.html' %} {% block page_title %} Popularity Based Song
 Recommendations {% endblock %} {% block body_block %}
 <section class="contact-form-wrap section">
   <div class="container-fluid">
     <div class="row p-5">
       <div class="col-lg-12 col-md-12 col-sm-12 title-heading-justify">
-        <h2 class="mb-4 heading-align">InfoBeans : Mood Based Song</h2>
+        <h2 class="mb-4 heading-align">InfoBeans :Popularity Based Song</h2>
         <h2>Recommendations</h2>
       </div>
 
@@ -14,7 +14,7 @@ Recommendations {% endblock %} {% block body_block %}
           id="contact-form"
           class="contact__form"
           method="post"
-          action="/mood-based-recommendation"
+          action="/songs/popularity-based-recommendation"
         >
           <!-- form message -->
 
@@ -60,9 +60,7 @@ Recommendations {% endblock %} {% block body_block %}
             rows="10"
             placeholder=""
             disabled
-          >
-{{common_mood}}</textarea
-          >
+          ></textarea>
         </div>
       </div>
     </div>
@@ -80,6 +78,7 @@ Recommendations {% endblock %} {% block body_block %}
     document.getElementById("msgarea2").value = "";
   }
 </script>
+
 <script>
   // Get the textarea element
   var textarea = document.getElementById('msgarea2');
@@ -88,13 +87,13 @@ Recommendations {% endblock %} {% block body_block %}
   var content = "";
 
   {% if songname %}
-  content += "Mood of  your song {{songname}}: {{common_mood}}\n";
+  content += "Popularity of your song {{songname}}: {{song_predicted_rank}}\n";
   {% endif %}
 
   {% if similar_songs|length > 0 %}
   content += "Similar songs:\n";
-  {% for index, song in similar_songs.iterrows() %}
-  content += "{{ song['name'] }}\n";
+  {% for song in similar_songs %}
+  content += "{{ song }}\n";
   {% endfor %}
   {% endif %}
   content+="{{ message }}"
