+++ poc/songsdetection/trained_model_classifiers/Decision Tree Classifier.joblib
