+++ poc/songsdetection/trained_model_classifiers/LGBM Classifier.joblib
