+++ poc/songsdetection/trained_model_classifiers/Support Vector Classifier.joblib
