--- templates/spam-ham-detection.html
@@ -1,62 +1,0 @@
-{% extends 'layouts/base.html' %}
-
-{% block page_title %}
-    Spam Ham Detection
-{% endblock %}
-
-{% block body_block %}
-<section class="contact-form-wrap section">
-    <div class="container-fluid" >
-        <div class="row p-5">
-            <div class="col-lg-12 col-md-12 col-sm-12 title-heading-justify">
-                <h2 class="mb-4 heading-align">InfoBeans : Spam Detection Tool for
-                </h2>
-                <h2> Messages/Emails!</h2>
-            </div>
-
-            <div class="col-lg-5 col-md-12 col-sm-12">
-               <!--  {% include 'inc/error_mess.html' %} -->
-                 <form id="contact-form" class="contact__form" method="post" action="{{ url_for('investor') }}">
-                 <!-- form message -->
-
-                    <!-- end message -->
-                    <h5 class="">Write Your Text Here                  </h5>
-                    <div class="form-group-2 mb-4">
-                        <textarea id="msgarea" name="message" class="form-control" rows="10" placeholder="Type Anything...">{% if message %}{{message}}{% endif %} </textarea>
-                    </div>
-                    <button class="blue buttn" id="submit_query" name="submit" type="submit">Submit</button>
-                    <button class="blue buttn" type="button" onClick="eraseText();" >Clear<i class="ti-reload"></i></button>
-                    <div id="loading-icon" style="display: none;">
-                        <!-- Replace with your loading icon HTML or image -->
-                        Loading...
-                    </div>
-                </form>
-            </div>
-            <div class="col-lg-7 col-md-12 col-sm-12">
-                <!-- form message -->
-                   <!-- end message -->
-                   <h5 class="">Answer</h5>
-                   <div class="form-group-2 mb-4">
-                       <textarea disabled id="msgarea2" name="message_output" class="form-control" rows="10" placeholder="">{% if result %}{{result}}{% endif %}</textarea>
-                   </div>
-
-           </div>
-        </div>
-    </div>
-</section>
-{% endblock %}
-
-{% block body_scripts %}
-<!--
-#########################################################
-##########          Essential Scripts          ##########
-#########################################################
--->
-<script>
-    function eraseText() {
-    document.getElementById("msgarea").value = "";
-    document.getElementById("msgarea2").value = "";
-    }
-</script>
-
-{% endblock %}
